+++ CodeConverter/CSharp/VisualBasicConverter.cs
@@ -8,13 +8,15 @@ using Microsoft.CodeAnalysis.Formatting;
 using VBSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax;
 using CSS = Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Threading;
+using ICSharpCode.CodeConverter.VB;
 
 namespace ICSharpCode.CodeConverter.CSharp
 {
     internal static class VisualBasicConverter
     {
         public static async Task<SyntaxNode> ConvertCompilationTreeAsync(Document document,
-            CSharpCompilation csharpViewOfVbSymbols, Project csharpReferenceProject, CancellationToken cancellationToken)
+            CSharpCompilation csharpViewOfVbSymbols, Project csharpReferenceProject,
+            OptionalOperations optionalOperations, CancellationToken cancellationToken)
         {
             document = await document.WithExpandedRootAsync(cancellationToken);
             var root = await document.GetSyntaxRootAsync(cancellationToken) as VBSyntax.CompilationUnitSyntax ??
@@ -30,13 +32,7 @@ namespace ICSharpCode.CodeConverter.CSharp
                 DeclarationNodeVisitor(document, compilation, semanticModel, csharpViewOfVbSymbols, csSyntaxGenerator);
             var converted = (CSS.CompilationUnitSyntax)await root.AcceptAsync(visualBasicSyntaxVisitor.TriviaConvertingDeclarationVisitor);
 
-            try {
-                // This call is very expensive for large documents. Should look for a more performant version, e.g. Is NormalizeWhitespace good enough?
-                converted = (CSS.CompilationUnitSyntax)Formatter.Format(converted, document.Project.Solution.Workspace, cancellationToken: cancellationToken);
-                return LineTriviaMapper.MapSourceTriviaToTarget(root, converted);
-            } catch (Exception) { //TODO log
-                return converted;
-            }
+            return optionalOperations.MapSourceTriviaToTargetHandled(root, converted, document);
         }
 
         private static string NullRootError(Document document)