+++ CodeConverter/Shared/ConversionOptions.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.CodeAnalysis;
+﻿using System;
+using Microsoft.CodeAnalysis;
 
 namespace ICSharpCode.CodeConverter.Shared
 {
@@ -6,5 +7,6 @@ namespace ICSharpCode.CodeConverter.Shared
     {
         public string RootNamespaceOverride { get; set; }
         public CompilationOptions TargetCompilationOptionsOverride { get; set; }
+        public TimeSpan AbandonOptionalTasksAfter { get; set; } = TimeSpan.FromMinutes(30);
     }
 }